+++ packages/frontend/src/components/dashboard/index.tsx
@@ -69,8 +69,11 @@ const StyledFunctionBar = styled(FunctionBar)`
 const cardCss = css`
   width: 928px;
 
-  ${mq.tabletAndBelow`
-    width: 100%;
+  ${mq.tabletOnly`
+    width: calc( 100vw - 64px );
+  `}
+  ${mq.mobileOnly`
+    width: calc( 100vw - 48px);
   `}
 `
 const CardBox = styled.div`
@@ -154,18 +157,25 @@ const CardSection = styled.div<{
     scrollbar-width: none;
   `
       : ''}
-  ${(props) =>
-    props.$windowWidth
-      ? `
-    max-width: 100%;
-    padding: 0 ${(props.$windowWidth - 928) / 2}px;
-  `
-      : 'max-width: 928px;'}
+
+  ${mq.desktopAndAbove`
+    ${(props) =>
+      props.$windowWidth
+        ? `
+        max-width: 100%;
+        padding: 0 ${
+          props.$isSidebarOpened ? 0 : (props.$windowWidth - 928) / 2
+        }px 0 ${(props.$windowWidth - 928) / 2}px;
+      `
+        : `
+        max-width: 928px;
+      `}
+  `}
+
   ${(props) =>
     props.$isSidebarOpened
       ? `
       width: 100vw;
-      padding-right: 24px;
     `
       : ''}
 
@@ -254,12 +264,14 @@ const Dashboard = () => {
     const sidebarComponent = sidebarRefs.current[selectedType]
     const cardComponent = cardRef.current
     if (sidebarComponent && cardComponent) {
-      const needGap = sidebarComponent.clientWidth + 24
+      let needGap = sidebarComponent.offsetWidth
       const hasGap = cardComponent.offsetLeft
-      if (cardComponent.clientWidth === 928) {
+      if (cardComponent.offsetWidth === 928) {
         // desktop and above
+        needGap += 32
         newSidebarGap = hasGap > needGap ? 0 : needGap
       } else {
+        needGap += 24
         newSidebarGap = needGap - hasGap
       }
     }